3eme cour JavaScript

+++ Introduction à la programmation/Exercises/Exo10_6.java
@@ -1,6 +1,6 @@
 public class Exo10_6 {
-    public static long factorielle(int n) {
-        long fact = 1L;
+    public static int factorielle(int n) {
+        int fact = 1;
 
         for (int i = 1; i <= n; i++) {
             fact *= i;
